requirements.txt
Pillow>=10.0.0
numpy>=1.24.0
google-generativeai>=0.8.0
scipy>=1.11.0
